moved logger to managers
